lots of debugging printouts
Former-commit-id: 9c180cbc2366dbd85691ae18ce0aa2dbf808e5b5 [formerly 9d9b0e0aa7dd3a9c8913189471671c838eeea084] [formerly 78645835b74ab271351d12d099698c3b8f6067c9 [formerly 01db17807333b8e03a7dff26e0e4ab359869800a]]
Former-commit-id: dd046a71cebd3504ca16e02dc5a47482219af1bb [formerly 8b64b8435185a68ad76c0c5f7f5bc3f3de6cf135]
Former-commit-id: 0566ca7e7e16db6870a4670aa909bd58b1ab743e

+++ src/data.cc
@@ -174,9 +174,9 @@ namespace FT
             
             if(o->Z.size() > 0)
             {
-                std::vector<int> zidx(o->y.size());
+                std::vector<long> zidx(o->y.size());
                 std::iota(zidx.begin(), zidx.end(), 0);
-                Eigen::VectorXi zw = Map<VectorXi>(zidx.data(), zidx.size());
+                VectorXl zw = Map<VectorXl>(zidx.data(), zidx.size());
                 zw = (zw.transpose()*perm).transpose();       // shuffle zw too
                 zidx.assign(((int*)zw.data()), (((int*)zw.data())+zw.size()));
